feat: set CORS_ALLOWED_ORIGINS
